+++ src/scripts/about.js
@@ -1,13 +1,10 @@
 import '../pages/about.css';
-import {ApiGitHub} from "./apiGitHub.js";
-import {CommitList} from "./commitList.js";
-import {CommitCard} from "./commitCard.js";
+import {CommitList} from "./modules/commitList.js";
+import {apiGitHub, commitCard} from "./constants/constants.js"
 
 import '../../node_modules/swiper/js/swiper.min.js';
 import Swiper from 'swiper';
 
-const apiGitHub = new ApiGitHub('https://api.github.com/repos/Atichka/NewsAnalyzer/commits');
-const commitCard = new CommitCard();
 const commitsList = new CommitList(document.querySelector('.history__swiper-wrapper'), commitCard);
 
 setTimeout(function () {
