+++ src/scripts/modules/newsCard.js
